jobs/requirements.txt
# === Existing template dependencies ===
requests==2.31.0
pandas==2.2.0
numpy==1.26.3
psycopg2-binary==2.9.9
python-dotenv==1.0.1
yfinance==0.2.36
schedule==1.2.1
sqlalchemy==2.0.25
dagster
dagster-webserver

# === Stream pipeline dependencies ===
redis[hiredis]>=5.0.0
websockets>=12.0
clickhouse-connect>=0.7.0

# === AI engineering dependencies ===
scikit-learn>=1.4.0
